Update rotation keys and CKKS to use Galois automorphisms + slot operations (WIP)
✅ ROTATION KEYS:
- Updated to use HashMap<usize, (Vec, Vec)> mapping automorphism index → key pair
- Generate keys for rotations -7 to +7 (needed for 8-component multivectors)
- Use proper Galois automorphisms via rotation_to_automorphism()

✅ CKKS ROTATION:
- Rewritten rotate_slots() to use Galois automorphisms properly
- Converts rotation amount r → automorphism index k = 5^r mod M
- Apply σₖ to ciphertext components, then use rotation key for key switching

✅ SLOT OPERATIONS MODULE:
- Created slot_operations.rs with extract_slot(), place_at_slot()
- Strategy: rotate to position 0, mask, rotate back
- Helper functions for optimized operations

⚠️  TESTS FAILING:
- Slot operation tests show large errors (~10^6)
- Issue likely with masking operation or scale management
- Needs debugging of mask multiplication and rotation interaction

Next: Debug why masking isn't working correctly, possibly scale mismatch
or incorrect mask polynomial construction.

+++ src/clifford_fhe/operations.rs
@@ -71,9 +71,9 @@ pub fn compute_component_product(
 
     // Rotate to move product from product_position to target_position
     let rotation_amount = if target_position >= product_position {
-        target_position - product_position
+        (target_position - product_position) as isize
     } else {
-        params.n + target_position - product_position
+        (params.n + target_position - product_position) as isize
     };
 
     if rotation_amount == 0 {
